+++ common/src/main/java/com/faboslav/featurify/common/versions/VersionedId.java
@@ -5,7 +5,7 @@ import net.minecraft.resources.ResourceKey;
 
 public class VersionedId
 {
-	public static Identifier GetId(ResourceKey<?> resourceKey) {
+	public static Identifier getId(ResourceKey<?> resourceKey) {
 		return resourceKey/*? if >= 1.21.11 {*/.identifier()/*?} else {*//*.location()*//*?}*/;
 	}
 }