+++ fabric/src/main/java/com/faboslav/featurify/fabric/FeaturifyFabricServer.java
@@ -1,22 +1,15 @@
 package com.faboslav.featurify.fabric;
 
-import com.faboslav.featurify.common.FeaturifyClient;
 import com.faboslav.featurify.common.events.common.LoadConfigEvent;
 import com.faboslav.featurify.common.events.common.UpdateWorldgenDataEvent;
 import com.faboslav.featurify.common.registry.RegistryManagerProvider;
-import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
+import net.fabricmc.api.DedicatedServerModInitializer;
 
-public final class FeaturifyFabricClient implements ClientModInitializer
+public final class FeaturifyFabricServer implements DedicatedServerModInitializer
 {
 	@Override
-	@Environment(EnvType.CLIENT)
-	public void onInitializeClient() {
-		FeaturifyClient.init();
-
+	public void onInitializeServer() {
 		LoadConfigEvent.EVENT.invoke(new LoadConfigEvent());
 		UpdateWorldgenDataEvent.EVENT.invoke(new UpdateWorldgenDataEvent(RegistryManagerProvider.getRegistryManager()));
 	}
 }
-