+++ neoforge/src/main/java/com/faboslav/featurify/neoforge/FeaturifyNeoForge.java
@@ -14,6 +14,7 @@ import net.neoforged.bus.api.EventPriority;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.neoforged.fml.loading.FMLEnvironment;
 import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.neoforge.event.RegisterCommandsEvent;
@@ -41,6 +42,7 @@ public final class FeaturifyNeoForge
 		}
 
 		modEventBus.addListener(FeaturifyNeoForge::onRegisterPayloadHandlers);
+		modEventBus.addListener(FeaturifyNeoForge::onCommonSetup);
 
 		eventBus.addListener(FeaturifyNeoForge::registerCommand);
 		eventBus.addListener(EventPriority.LOWEST, FeaturifyNeoForge::onResourceManagerReload);
@@ -49,6 +51,13 @@ public final class FeaturifyNeoForge
 
 	private static void onRegisterPayloadHandlers(final RegisterPayloadHandlersEvent event) {
 		NeoForgePlatformNetwork.onRegisterPayloadHandlers(event);
+	}
+
+	private static void onCommonSetup(final FMLCommonSetupEvent event) {
+		event.enqueueWork(() -> {
+			LoadConfigEvent.EVENT.invoke(new LoadConfigEvent());
+			UpdateWorldgenDataEvent.EVENT.invoke(new UpdateWorldgenDataEvent(RegistryManagerProvider.getRegistryManager()));
+		});
 	}
 
 	private static void registerCommand(RegisterCommandsEvent event) {